requirements.txt
flask>=2.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
requests>=2.31.0
